pyproject.toml
[tool.poetry]
name = "dev"
version = "0.1.0"
description = ""
authors = ["woohaen88 <woohaen88@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
black = "^23.7.0"
pillow = "^10.0.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
